refactor: move increment counter to ui
